move pre-generated code out of code-fmt scope
