+++ benchmark/pom.xml
@@ -66,6 +66,26 @@
           </execution>
         </executions>
       </plugin>
+      <plugin>
+        <artifactId>maven-resources-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>copy-pregen</id>
+            <phase>generate-sources</phase>
+            <goals>
+              <goal>copy-resources</goal>
+            </goals>
+            <configuration>
+              <outputDirectory>target/generated-sources/jflex/de/jflex/benchmark/pregen</outputDirectory>
+              <resources>
+                <resource>
+                  <directory>src/main/pregen</directory>
+                </resource>
+              </resources>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
       <!-- Build an Uberjar -->
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
