+++ benchmark/src/main/pregen/NoAction17.java
@@ -1,6 +1,6 @@
 /* The following code was generated by JFlex 1.7.0 */
 
-package de.jflex.benchmark;
+package de.jflex.benchmark.pregen;
 
 /*
   A scanner with minimal action code, to measure inner matching loop
